SpatialManning (spelt correctly) updated
[x] SpatialMapping namespace updated
[x] SpacialMapping editor namespace updated
