+++ Assets/MixedRealityToolkit/SpatialMapping/Scripts/RemoteMapping/Editor/FileSurfaceObserverEditor.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 374438d4d75acaf4ab87c1d3e97a6e55
+timeCreated: 1515951385
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
