+++ Assets/MixedRealityToolkit/SpatialMapping/Scripts/RemoteMapping/Editor/RoomMeshExporter.cs
@@ -1,9 +1,12 @@
-﻿using System.Collections.Generic;
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+using MixedRealityToolkit.Common.Editor;
+using MixedRealityToolkit.SpatialMapping.RemoteMapping;
+using System.Collections.Generic;
 using System.IO;
-using System.Text;
 using UnityEditor;
 using UnityEngine;
-using HoloToolkit.Unity;
 
 namespace HoloToolkit.Unity.SpatialMapping
 {
