+++ Assets/MixedRealityToolkit/SpatialMapping/Scripts/RemoteMapping/MeshSaver.cs
@@ -12,7 +12,7 @@ using Windows.Storage;
 using Windows.Storage.Streams;
 #endif
 
-namespace HoloToolkit.Unity.SpatialMapping
+namespace MixedRealityToolkit.SpatialMapping.RemoteMapping
 {
     /// <summary>
     /// MeshSaver is a static class containing methods used for saving and loading meshes.
